requirements.txt
aioice==0.9.0
aiortc==1.10.1
annotated-types==0.7.0
anyio==4.8.0
av==13.1.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.11
google-crc32c==1.6.0
h11==0.14.0
idna==3.10
ifaddr==0.2.0
numpy==2.2.3
opencv-python==4.11.0.86
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
pylibsrtp==0.11.0
pyOpenSSL==25.0.0
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
